backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Akram"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
fastapi-utils = "^0.8.0"
uvicorn = "^0.34.2"
gunicorn = "^23.0.0"
python-dotenv = "^0.20.0"
mongoengine = "^0.24.1"
loguru = "^0.6.0"
requests = "^2.28.1"
anyio = "^4.4.0"
httpx = "^0.27.0"
pytest-anyio = "^0.0.0"
pytest-mongodb = "^2.4.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-multipart = "^0.0.9"
langchain = "^0.3.25"
langchain-ollama = "^0.3.2"
langchain-chroma = "^0.2.3"
pysqlite3-binary = "^0.5.4"
pydantic = {extras = ["email"], version = "^2.11.4"}

[poetry.group.dev.dependencies]
essential-generators = "^1.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.3"
asgi-lifespan = "^1.0.1"
coverage = "^6.4.1"
coverage-badge = "1.1.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
